requirements.txt
# plugins for the website
django-registration
south